move two controllers to stimulus
